+++ custom-elements/auto-complete-element/spec/index.spec.ts
@@ -27,9 +27,15 @@ describe('AutoCompleteElement', () => {
       options = list.querySelectorAll('[role="option"]');
     });
 
+    it('#multiple returns false', () => {
+      expect(el.multiple).to.equal(false);
+    });
+
     it('sets the default attributes', () => {
       expect(input).to.have.attribute('spellcheck', 'false');
       expect(input).to.have.attribute('autocomplete', 'off');
+      expect(list).to.have.attribute('tabindex', '-1');
+      expect(list).to.have.attribute('aria-orientation', 'vertical');
     });
 
     it('opens the menu on focus', () => {
@@ -359,6 +365,10 @@ describe('AutoCompleteElement', () => {
       input = el.querySelector('input');
       list = el.querySelector('ul');
       options = list.querySelectorAll('[role="option"]');
+    });
+
+    it('#multiple returns true', () => {
+      expect(el.multiple).to.equal(true);
     });
 
     it('does not close the list and does not update the input value', () => {